requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
python-docx>=1.0.0
SQLAlchemy>=2.0.0
openai>=1.12.0
python-dotenv>=1.0.0
